refactor detection and flyblock validators to use a common base class
